% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{accessors}
\alias{accessors}
\alias{getBanterModel}
\title{BANTER class accessor functions}
\usage{
getBanterModel(x, model = "event")
}
\arguments{
\item{x}{a \code{\link{banter_model}} object.}

\item{model}{model to retrieve.}
}
\description{
Accessor functions for components of BANTER models
}
