% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addBanterDetector.R
\name{addBanterDetector}
\alias{addBanterDetector}
\alias{removeBanterDetector}
\title{Add a BANTER Detector Model}
\usage{
addBanterDetector(x, data, ntree, sampsize = 1, num.cores = NULL)

removeBanterDetector(x, name)
}
\arguments{
\item{x}{a \code{\link{banter_model}} object.}

\item{data}{detector data.frame.}

\item{ntree}{number of trees.}

\item{sampsize}{number or fraction of samples to use in each tree.}

\item{num.cores}{number of cores to use for Random Forest model.}

\item{name}{name of detector.}
}
\value{
a \code{\link{banter_model}} object with the detector model added or 
  removed.
}
\description{
Add a detector model to a BANTER classifier.
}
\examples{
data(train.data)
# initialize BANTER model with event data
bant.mdl <- initBanterModel(train.data$events)
# add the 'bp' (burst pulse) detector model
bant.mdl <- addBanterDetector(
  bant.mdl, train.data$detectors$bp, 
  ntree = 50, sampsize = 1, num.cores = 1
)
bant.mdl

# remove the 'bp' detector model
bant.mdl <- removeBanterDetector(bant.mdl, "bp")
bant.mdl

}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
