% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addDetectorModel.R
\name{addDetectorModel}
\alias{addDetectorModel}
\alias{removeDetectorModel}
\title{Add a Detector Model}
\usage{
addDetectorModel(x, name, data, ntree, sampsize = 1)

removeDetectorModel(x, name)
}
\arguments{
\item{x}{a \code{\link{event_model}} object.}

\item{name}{name of detector}

\item{data}{detector data.frame}

\item{ntree}{number of trees}

\item{sampsize}{number or fraction of samples to use in each tree}
}
\description{
Add a detector model to an event model
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
