% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/banter-package.R
\name{banter-package}
\alias{banter-package}
\alias{banter}
\title{BioAcoustic EveNT ClassifiER}
\description{
banter
}
\references{
Rankin, S. , Archer, F. , Keating, J. L., Oswald, J. N., 
  Oswald, M. , Curtis, A. and Barlow, J. (2017), Acoustic classification 
  of dolphins in the California Current using whistles, echolocation clicks,
  and burst pulses. Marine Mammal Science 33:520-540. doi:10.1111/mms.12381
}
\keyword{package}
