% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/banter_detector.R
\docType{class}
\name{banter_detector-class}
\alias{banter_detector-class}
\alias{banter_detector}
\title{Detector classification model}
\arguments{
\item{name}{detector name}

\item{detector.data}{measurement data for detections}

\item{event.data}{data.frame associating event.id with species}

\item{ntree}{number of trees in model}

\item{sampsize}{number or fraction of samples to use in each tree}
}
\description{
Create classification model for a detector
}
\section{Slots}{

\describe{
\item{\code{name}}{detector name}

\item{\code{ids}}{data.frame of event.ids and call.ids for calls in detector}

\item{\code{model}}{classification model}
}}

\author{
Eric Archer \email{eric.archer@noaa.gov}
}
\keyword{internal}
