% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/banterGuide.R
\name{banterGuide}
\alias{banterGuide}
\title{BANTER Guide}
\usage{
banterGuide()
}
\description{
Open a browser window displaying 
  "BANTER: A User’s Guide to Acoustic Classification".
}
\references{
Rankin, S. , Archer, F. , Keating, J. L., Oswald, J. N., 
  Oswald, M. , Curtis, A. and Barlow, J. (2017), Acoustic classification 
  of dolphins in the California Current using whistles, echolocation clicks,
  and burst pulses. Marine Mammal Science 33:520-540. doi:10.1111/mms.12381
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
