% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildEventModel.R
\name{buildEventModel}
\alias{buildEventModel}
\title{Build Event Model}
\usage{
buildEventModel(x, ntree, sampsize = 1)
}
\arguments{
\item{x}{a \code{\link{event_model}} object.}

\item{ntree}{number of trees}

\item{sampsize}{number or fraction of samples to use in each tree}
}
\description{
Build full event classifier model
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
