% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{classes}
\alias{classes}
\alias{detector_model-class}
\alias{detector_model}
\title{detector_model class}
\description{
An S4 class storing a classification model for a detector.
}
\section{Slots}{

\describe{
\item{\code{name}}{detector name}

\item{\code{model}}{classification model}
}}

\author{
Eric Archer \email{eric.archer@noaa.gov}
}
