% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_model.R
\docType{class}
\name{event_model-class}
\alias{event_model-class}
\alias{event_model}
\title{Event classification model}
\arguments{
\item{detector.models}{measurement data for detections}

\item{event.data}{data.frame associating event.id with species}

\item{ntree}{number of trees in model}

\item{sampsize}{number or fraction of samples to use in each tree}
}
\description{
Create classification model for events
}
\section{Slots}{

\describe{
\item{\code{data}}{data.frame of event.ids and call.ids for calls in detector}

\item{\code{detectors}}{list of \code{detector_model} objects}

\item{\code{model}}{classification model}
}}

\author{
Eric Archer \email{eric.archer@noaa.gov}
}
