% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBanterModel.R
\name{getBanterModel}
\alias{getBanterModel}
\title{Extract Random Forest Model}
\usage{
getBanterModel(x, model = "event")
}
\arguments{
\item{x}{a \code{\link{banter_model}} object.}

\item{model}{name of model to extract. Default is \code{"event"} 
to summarize the event-level model. Can also be name of a detector.}
}
\value{
a \code{\link{randomForest}} model object.
}
\description{
Extract BANTER event or detector Random Forest model.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
