% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBanterModelData.R
\name{getBanterModelData}
\alias{getBanterModelData}
\title{Extract Random Forest Model Data}
\usage{
getBanterModelData(x)
}
\arguments{
\item{x}{a \code{\link{banter_model}} object.}
}
\value{
a \code{\link{randomForest}} model object.
}
\description{
Extract BANTER event data used for the Random Forest model.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
