% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDetectorNames.R
\name{getDetectorNames}
\alias{getDetectorNames}
\title{Detector Names}
\usage{
getDetectorNames(x)
}
\arguments{
\item{x}{a \code{\link{banter_model}} object.}
}
\value{
a vector of names.
}
\description{
Return names of detectors loaded in BANTER model.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
