% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSampSize.R
\name{getSampSize}
\alias{getSampSize}
\title{Sample Size}
\usage{
getSampSize(x, model = "event")
}
\arguments{
\item{x}{a \code{\link{banter_model}} object.}

\item{model}{name of model to extract. Default is \code{"event"} 
to summarize the event-level model. Can also be name of a detector.}
}
\value{
a vector of sample sizes.
}
\description{
Return sample size used for a BANTER model.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
