% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initBanterModel.R
\name{initBanterModel}
\alias{initBanterModel}
\title{Initialize BANTER model}
\usage{
initBanterModel(x)
}
\arguments{
\item{x}{a data.frame}
}
\description{
Initialize a BANTER model with a data.frame of events.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
