% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadDetectorModel.R
\name{loadDetectorModel}
\alias{loadDetectorModel}
\title{Detector classification model}
\usage{
loadDetectorModel(name, detector.data, event.species, ntree, samp.pct = 0.5,
  min.samp.n = 1)
}
\arguments{
\item{name}{detector name}

\item{detector.data}{measurement data for detections}

\item{event.species}{data.frame associating event.id with species}

\item{ntree}{number of trees in model}

\item{samp.pct}{fraction of samples to use in each tree}

\item{min.samp.n}{minimum number of samples per species for each tree}
}
\description{
Create classification model for a detector
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
