% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelDiagnostics.R
\name{modelDiagnostics}
\alias{modelDiagnostics}
\title{Random Forest Model Diagnostics}
\usage{
modelDiagnostics(x, detector = NULL, n = 0.1)
}
\arguments{
\item{x}{a \code{\link{event_model}} object.}

\item{detector}{name of detector. If \code{NULL}, returns event model
summary.}

\item{n}{number of final iterations to summarize OOB error rate for. If 
between 0 and 1 is taken as a proportion of chain.}
}
\description{
display diagnostics for event and detector models
}
