% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelPctCorrect.R
\name{modelPctCorrect}
\alias{modelPctCorrect}
\title{Model Percent Correct}
\usage{
modelPctCorrect(x)
}
\arguments{
\item{x}{a \code{\link{banter_model}} object.}
}
\value{
a data.frame with the percent correctly classified for each model 
  in \code{x}.
}
\description{
Extract percent correctly classified by species for detector 
  and event models.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
