% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numCalls.R
\name{numCalls}
\alias{numCalls}
\alias{propCalls}
\title{Number and Proportion of Calls}
\usage{
numCalls(x, by = c("species", "event"))

propCalls(x, by = c("species", "event"))
}
\arguments{
\item{x}{a \code{\link{banter_model}} object.}

\item{by}{return summary by \code{"species"} or \code{"event"}.}
}
\description{
Number and proportion of calls in BANTER detector models
}
