% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numEvents.R
\name{numEvents}
\alias{numEvents}
\title{Number of Events}
\usage{
numEvents(x, model = "event")
}
\arguments{
\item{x}{a \code{\link{banter_model}} object.}

\item{model}{name of model to extract. Default is \code{"event"} 
to summarize the event-level model. Can also be name of a detector.}
}
\value{
a data.frame giving the number of events available for each species.
}
\description{
Number of events in BANTER model by species.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
