% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict}
\alias{predict}
\alias{predict}
\alias{predict.event_model}
\alias{predict}
\alias{predict,event_model-method}
\title{Predict event data}
\usage{
predict(object, ...)

\method{predict}{event_model}(object, new.data, ...)
}
\arguments{
\item{object}{a \code{\link{event_model}} object.}

\item{...}{unused.}

\item{new.data}{a list of event and detector data.}
}
\description{
Predict event data
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
