% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runBanterModel.R
\name{runBanterModel}
\alias{runBanterModel}
\title{Run BANTER Model}
\usage{
runBanterModel(x, ntree, sampsize = 1)
}
\arguments{
\item{x}{a \code{\link{banter_model}} object.}

\item{ntree}{number of trees}

\item{sampsize}{number or fraction of samples to use in each tree}
}
\description{
Build full event classifier model
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
