% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/banter-package.R
\docType{data}
\name{test.data}
\alias{test.data}
\title{Testing events and detectors}
\format{
list
}
\usage{
data(test.data)
}
\description{
A list of events and call data from detectors for testing 
  BANTER model
}
\keyword{datasets}
