% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/banter-package.R
\docType{data}
\name{train.data}
\alias{train.data}
\title{Training events and detectors}
\format{
list
}
\usage{
data(train.data)
}
\description{
A list of events and call data from detectors for training 
  BANTER model
}
\keyword{datasets}
