% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{accessors}
\alias{accessors}
\alias{getModel}
\title{BANTER class accessor functions}
\usage{
getModel(x, model = "event")
}
\arguments{
\item{x}{a \code{\link{event_model}} object.}

\item{model}{model to retrieve.}
}
\description{
Accessor functions for slots in BANTER classes
}
