% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addBanterDetector.R
\name{addBanterDetector}
\alias{addBanterDetector}
\alias{removeBanterDetector}
\title{Add a BANTER Detector Model}
\usage{
addBanterDetector(x, data, ntree, sampsize = 1)

removeBanterDetector(x, name)
}
\arguments{
\item{x}{a \code{\link{banter_model}} object.}

\item{data}{detector data.frame}

\item{ntree}{number of trees}

\item{sampsize}{number or fraction of samples to use in each tree}

\item{name}{name of detector}
}
\description{
Add a detector model to a BANTER classifier.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
