% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/banter_detector.R
\docType{class}
\name{banter_detector-class}
\alias{banter_detector-class}
\alias{banter_detector}
\title{Detector classification model}
\description{
Create classification model for a detector
}
\section{Slots}{

\describe{
\item{\code{name}}{detector name}

\item{\code{ids}}{data.frame of event.ids and call.ids for calls in detector}

\item{\code{model}}{classification model}

\item{\code{sampsize}}{sample size vector used in model}

\item{\code{timestamp}}{start and stop times of model run}
}}

\author{
Eric Archer \email{eric.archer@noaa.gov}
}
\keyword{internal}
