% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/banter_model.R
\docType{class}
\name{banter_model-class}
\alias{banter_model-class}
\alias{banter_model}
\title{Events classification model}
\description{
Create classification model for events
}
\section{Slots}{

\describe{
\item{\code{data}}{data.frame of event.ids and call.ids for calls in detector}

\item{\code{detectors}}{list of \code{detector_model} objects}

\item{\code{model.data}}{data used to create classification model}

\item{\code{model}}{classification model}

\item{\code{sampsize}}{sample size vector used in model}

\item{\code{timestamp}}{start and stop times of model run}
}}

\author{
Eric Archer \email{eric.archer@noaa.gov}
}
\keyword{internal}
