% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSampsize.R
\name{getSampsize}
\alias{getSampsize}
\title{Get Balanced Sample Size}
\usage{
getSampsize(x, req.sampsize)
}
\arguments{
\item{x}{vector of discrete class values or table of frequencies as output 
of \code{\link{table}}.}

\item{req.sampsize}{desired number of samples per class. If value is between 
0 and 1, then it is treated as a percentage and sample size will be this 
value times the smallest class frequency rounded to the nearest whole 
number.}
}
\description{
Produces a vector of equal sample sizes for balanced 
  Random Forest model.
}
\details{
\code{req.sampsize} must be smaller than smallest class frequency. If
  \code{req.sampsize} is a percentage, then resulting rounded sample size 
  must be at least 1.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
