% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSampsize.R
\name{getSampsize}
\alias{getSampsize}
\title{Get Balanced Sample Size}
\usage{
getSampsize(x, samp.pct, min.samp.n = 1)
}
\arguments{
\item{x}{vector of classes as response values or table of frequencies 
(output of \code{\link{table}}.}

\item{samp.pct}{desired percentage to use from smallest class (\code{0:1}).}

\item{min.samp.n}{minimum permitted sample size.}
}
\description{
return a vector of equal sample sizes for balanced Random Forest 
  model. Takes vector of discrete values or frequency table as well as 
  desired sample size percentage and minimum sample size.
}
\details{
Returns vector of \code{floor(samp.pct * smallest class frequency)}
 that is as long as there are unique classes in \code{x}. Will throw an error 
 if this number is < \code{min.samp.n}.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
