% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R, R/modelDiagnostics.R
\name{internals}
\alias{internals}
\alias{.getSampsize}
\alias{.meanVotes}
\alias{.numCalls}
\alias{.propCalls}
\alias{.rfErrorSmry}
\alias{.inBagSmry}
\title{Produces a vector of equal sample sizes for balanced Random Forest model.}
\usage{
.getSampsize(x, n, warn.label)

.meanVotes(x)

.numCalls(x)

.propCalls(x)

.rfErrorSmry(rf, n = 0.1)

.inBagSmry(rf)
}
\description{
Produces a vector of equal sample sizes for balanced Random Forest model.

Returns matrix of mean species vote percentage by event

Returns matrix of number of calls by each detector for each event

Returns matrix of number of calls by each detector for each event
}
\keyword{internal}
