% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addBanterDetector.R, R/internals.R, R/summary.R
\name{.rfFuncDetector}
\alias{.rfFuncDetector}
\alias{.checkModelName}
\alias{.stopIfNoDetectors}
\alias{internals}
\alias{.getSampsize}
\alias{.meanVotes}
\alias{.rfPctCorrectSmry}
\title{Detector randomForest function}
\usage{
.rfFuncDetector(params)

.checkModelName(x, model)

.stopIfNoDetectors(x)

.getSampsize(x, n, warn.label)

.meanVotes(x)

.rfPctCorrectSmry(rf, n = 0.1)
}
\description{
Detector randomForest function

Checks to make sure 'model' is a valid model name

Produces a vector of equal sample sizes for balanced Random Forest model.

Returns matrix of mean species vote percentage by event
}
\keyword{internal}
