% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R, R/summary.R
\name{.checkModelName}
\alias{.checkModelName}
\alias{.stopIfNoDetectors}
\alias{internals}
\alias{.getSampsize}
\alias{.meanVotes}
\alias{.rfErrorSmry}
\alias{.inBagSmry}
\title{Checks to make sure 'model' is a valid model name}
\usage{
.checkModelName(x, model)

.stopIfNoDetectors(x)

.getSampsize(x, n, warn.label)

.meanVotes(x)

.rfErrorSmry(rf, n = 0.1)

.inBagSmry(rf)
}
\description{
Checks to make sure 'model' is a valid model name

Produces a vector of equal sample sizes for balanced Random Forest model.

Returns matrix of mean species vote percentage by event
}
\keyword{internal}
