% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict}
\alias{predict}
\alias{predict}
\alias{predict.banter_model}
\alias{predict}
\alias{predict,banter_model-method}
\title{Predict BANTER events}
\usage{
predict(object, ...)

\method{predict}{banter_model}(object, new.data, ...)
}
\arguments{
\item{object}{a \code{\link{banter_model}} object.}

\item{...}{unused.}

\item{new.data}{a list of event and detector data that has the same 
predictors as in the \code{banter_model}. It must contain elements called 
\code{events} and \code{detectors}. The \code{events} element must be a 
data.frame that has a column called \code{event.id} and the same
predictor columns as the event data used to initialize the banter model 
(see \code{\link{initBanterModel}}). The \code{detectors} element must be 
a named list with the same detectors used to build the model 
(see \code{\link{addBanterDetector}}).}
}
\value{
A list with the following elements: \describe{
  \item{events}{the data frame used in the event model for predictions.}
  \item{predict.df}{data.frame of predicted species and assignment 
     probabilities for each event.}
}
}
\description{
Predict species of events for novel data from a BANTER model.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
