% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict}
\alias{predict}
\alias{predict}
\alias{predict.banter_model}
\alias{predict}
\alias{predict,banter_model-method}
\title{Predict BANTER events}
\usage{
predict(object, ...)

\method{predict}{banter_model}(object, new.data, ...)
}
\arguments{
\item{object}{a \code{\link{banter_model}} object.}

\item{...}{unused.}

\item{new.data}{a list of event and detector data.}
}
\value{
A data.frame with the predicted species (\code{"predicted"}) followed 
  by columns of assignment probabilities.
}
\description{
Predict species of events for novel data from a BANTER model.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
