% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary}
\alias{summary}
\alias{summary}
\alias{summary.banter_model}
\alias{summary}
\alias{summary,banter_model-method}
\title{BANTER Classifier Model Summary}
\usage{
summary(object, ...)

\method{summary}{banter_model}(object, detector = "event", n = 0.1, ...)
}
\arguments{
\item{object}{a \code{\link{banter_model}} object.}

\item{...}{ignored.}

\item{detector}{name of detector to summarize. Default is \code{"event"} 
to summarize the event-level model.}

\item{n}{number of final iterations to summarize OOB error rate for. If 
between 0 and 1 is taken as a proportion of chain.}
}
\description{
Display summaries for event and detector models
}
