% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary}
\alias{summary}
\alias{summary}
\alias{summary.banter_model}
\alias{summary}
\alias{summary,banter_model-method}
\title{BANTER Classifier Model Summary}
\usage{
summary(object, ...)

\method{summary}{banter_model}(object, model = "event", n = 0.1,
  bins = 20, ...)
}
\arguments{
\item{object}{a \code{\link{banter_model}} object.}

\item{...}{ignored.}

\item{model}{name of model to summarize. Default is \code{"event"} 
to summarize the event-level model. Can also be name of a detector.}

\item{n}{number of final iterations to summarize OOB error rate for. If 
between 0 and 1 is taken as a proportion of chain.}

\item{bins}{number of bins in inbag histogram.}
}
\description{
Display summaries for event and detector models
}
